steampunk_hamclock/deploy_staging/steampunk_hamclock/backend/requirements.txt
fastapi
uvicorn
pyserial
pynmea2
requests
python-multipart
